\name{Ensemble.logit}
\alias{Ensemble.logit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Ensemble BMA Logit Function for Binary Outcomes
}
\description{
Creates an ensemble BMA model for dichotomous outcomes.  
} 
\usage{
Ensemble.logit(y, pp.raw, exp = 3, tol = 0.001, max.iter = 10000, model.names = NULL)
}
\arguments{
  \item{y}{
The observed dichotomous outcome variable. NOTE: Only the in-sample forecasts should be included.
}
  \item{pp.raw}{
An n by k matrix of predicted probabilies, where n is the number of observations in the in-sample dataset and k is the number of component forecasts.  Each column should contain n predicted probabilities.
}
  \item{exp}{
The exponential shrinkage term.  Forecasts are raised to the (1/exp) power on the logit scale for the purposes of bias reduction.  The default value is \code{exp=3}.
}
  \item{tol}{
Tolerance for improvements in the log-likelihood before the EM algorithm will stop optimization.  The default is \code{tol= 0.01}, which is somewhat high.  Researchers may wish to reduce this by an order of magnitude for final model estimation.
}
  \item{max.iter}{
The maximum number of iterations the EM algorithm will run before stopping automatically.
}
  \item{model.names}{
  A vector containing the labels for each component model to be included in tables and plots.  Default value is NULL, which simply numbers each model or uses the column names of \code{pp.raw} if provided.
}
}
\details{
Given a binary dependent variable and in-sample predictions from multiple component forecast models, \code{Ensemble.logit} fits an ensemble BMA mixture model for dichotomous outcomes. The weights assigned to each model are derived from the individual model's performance on the in-sample outcomes (y). 
}
\value{
  \item{W}{The posterior weights assigned to each model}
  \item{log.lik}{The final log-likelihood for the EBMA model}
  \item{pred.prob}{The posterior predicted probabilities for all n in-sample observations generated by the ensemble model}
  \item{model.params}{The bias-reduction constant and predictor coeficients estimated for each of the k models}
  \item{exp}{The exponential shrinkage term specified by the user}
  \item{model.names}{A string vector containing the name of each component model}
  \item{model.pps}{A n by k matrix containing the predicted probabilities from each component model (the same matrix as \code{pp.raw} above)}
  \item{bma.brier}{The brier score for the in-sample EBMA model}
  \item{models.brier}{The brier score for the component model forecasts}
  \item{bma.perc}{The percent of observations correctly predicted by the in-sample EBMA model}
  \item{models.perc}{The percent of observations correctly predicted by the component model forecasts}
  \item{bma.pre}{The proportional reduction in error for the in-sample EBMA model, where the baseline model always predicts zero.}
  \item{models.pre}{The proportional reduction in error for the component models, where the baseline model always predicts zero.}
  \item{bma.auc}{The area under the ROC curve for the in-sample EBMA model}
  \item{models.auc}{The area under the ROC curve for the component model forecasts}
  \item{bma.sens}{The sensitivity statistic for the in-sample EBMA model}
  \item{models.sens}{The sensitivity statistic for the component models}
  \item{bma.spec}{The specificity statistic for the in-sample EBMA model}
  \item{models.spec}{The specificity statistic for the component models}
  \item{dv}{The actual observed values for the outcome variable for the in-sample dataset (the same as \code{y} above)}
  \item{n}{The number of complete observations in the in-sample dataset}
}
\author{
Michael D. Ward and Jacob Montgomery <jacob.montgomery@duke.edu>
}
\references{
Ward, M. D., J. M. Montgomery and F. M. Hollenbach. 2011 Improving Conflict Predictions Using Ensemble Bayesian Model Averaging.  Paper presented at the 2011 meeting of the International Studies Association in Montreal, Quebec.\cr
\cr
Raftery, A. E., T. Gneiting, F. Balabdaoui and M. Polakowski. 2005 Using Bayesian Model Averaging to calibrate forecast ensembles, \emph{Monthly Weather Review} \bold{133}:1155--1174.\cr
\cr
Sloughter, J. M., A. E. Raftery, T. Gneiting and C. Fraley. 2007 Probabilistic quantitative precipitation forecasting using Bayesian model averaging, \emph{Monthly Weather Review} \bold{135}:3209--3220.\cr
\cr
Fraley, C., A. E. Raftery, T. Gneiting, 2010 Calibrating Multi-Model Forecast Ensembles with Exchangeable and Missing Members using Bayesian Model Averaging, \emph{Monthly Weather Review} \bold{138}:190--202.\cr
\cr
Sloughter, J. M., T. Gneiting and A. E. Raftery, 2010 Probabilistic wind speed forecasting using ensembles and Bayesian model averaging, \emph{Journal of the American Statistical Association}, \bold{105}:25--35.
}
\keyword{ package }
\seealso{
\code{\link{EBMAforecast}}, \code{\link{predict.Ensemble.logit}}, \code{\link{Insample}}, \code{\link{Outsample}}
}
\note{
PLEASE NOTE: This function does not currently handle missing data, but does not provide errors.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
demo(EBMAforecast)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{BMA}

