\name{Outsample}
\alias{Outsample}
\docType{data}
\title{Out-of-Sample Predictions from the Seven Forecasting Models
}
\description{
Outsample provides the out-of-sample predictions of seven different models predicting insurgency in 29 Asian countries, as well as an indicator variable for the true occurence of insurgency. The out-of-sample data ranges from January 2009 to December 2010.
}
\usage{data(Outsample)}
\format{
  A data frame with 696 observations on the following 11 variables.
  \describe{
    \item{\code{Country}}{A factor with levels \code{AUSTRALIA} \code{BANGLADESH} \code{BHUTAN} \code{CAMBODIA} \code{CHINA} \code{COMOROS} \code{FIJI} \code{INDIA} \code{INDONESIA} \code{JAPAN} \code{LAOS} \code{MADAGASCAR} \code{MALAYSIA} \code{MAURITIUS} \code{MONGOLIA} \code{MYANMAR} \code{NEPAL} \code{NEW ZEALAND} \code{NORTH KOREA} \code{PAPUA NEW GUINEA} \code{PHILIPPINES} \code{RUSSIA} \code{SINGAPORE} \code{SOLOMIN ISLANDS} \code{SOUTH KOREA} \code{SRI LANKA} \code{TAIWAN} \code{THAILAND} \code{VIETNAM}}
    \item{\code{Year}}{A numeric vector of the years in the out-of-sample data}
    \item{\code{Month}}{A numeric vector of the months in the out-of-sample data}
       \item{\code{Polisci}}{A numeric vector of the out-of-sample predictions of the Polisci model in the ICEWS example. The polisci model is a hierachical model that includes a number of political variables from the Polity IV dataset and other political variables as covariates, as well as country specific intercepts.}
    \item{\code{Glm.Econ}}{A numeric vector of the out-of-sample predictions of the Glm.econ model in the ICEWS example. These are the predictions generated by a simple GLM model with gdp growth, number of minotriy groups at risk, population size and a count of cooperative events with the USA as independent variables.}
    \item{\code{Lmer3}}{A numeric vector of the out-of-sample predictions of the Lmer3 model in the ICEWS example. The Lmer3 model is similar to the Polisci model above as it includes several different political variables (such as executive constraints or competitiveness of participation) as covariates, country specific intercepts, and a random country-level coefficient for per capita GDP. In addition a spatial lag of conflictual events in neighboring countries is included.}
    \item{\code{Glm.simple}}{A numeric vector of the out-of-sample predictions of the Glm.simple model in the ICEWS example.The GLM.Simple is the simplest model included in the example dataset. It only includes population size and gdp growth as independent variables. The Glm.simple model generally performs the worst of the component models.}
    \item{\code{Lmer.all}}{A numeric vector of the out-of-sample predictions of the Lmer.all model in the ICEWS example. This model is very similar to the Polisci and Lmer3 model as various political variables, such as competitiveness of participation or occurence of anti-government violence are used as covariates and country specific intercepts are also included.}
    \item{\code{Icews}}{A numeric vector of the out-of-sample predictions of the Duke model in the ICEWS example. The Duke model was developed by a team at Duke University to forecast insurgency events as part of the ICEWS project. The Duke model is a hierarchical model with several variables (such as number of conflictual events that involves the country's military and time since last election) included as covariates, while a spatial lag for the average number of domestic political crises in neighboring countries and a country specific effect for gdp per capita are also included. In general this was the most accurate of the component models.}
    \item{\code{Insurgency}}{A numeric vector of the true values of insurgency occurence in the in-sample data set}
    \item{\code{SAE}}{A numeric vector of the out-of-sample predictions of the SAE model in the ICEWS example. The SAE model was developed to produce forecasts of conflictual events as part of the ICEWS project and was designed by the SAE team. The model is a generalized linear model using 27 different independent variables. All of the variables are taken from basic event stream data provided through the ICEWS project. The SAE model was generally the second most accurate of the component models and comes very close to the Duke model in its performance.}
  }
}
\source{
Ward, M. D., J. M. Montgomery and F. M. Hollenbach. 2011 Improving Conflict Predictions Using Ensemble Bayesian Model Averaging.  Paper presented at the 2011 meeting of the International Studies Association in Montreal, Quebec.
}
\examples{
data(Outsample)
}
\keyword{datasets}
