\name{predict.Ensemble.logit}
\alias{predict.Ensemble.logit}
%- Also NEED an '\alias' for EACH other topic documented here.  
\title{
Generate Out-of-Sample Predictions Using Ensemble BMA Models for Binary Outcomes  
}
\description{Calculates out-of-sample forecasts on the predicted probability scale for the ensemble model.  Inputs include an \code{Ensemble.logit} object and out-of-sample predictions (on the predicted probability scale) generated from component forecasts.
}
\usage{
predict.Ensemble.logit(obj, newdata, y.out)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obj}{
An object created by the \code{\link{Ensemble.logit}} function
}
  \item{newdata}{
A matrix of out-of-sample predicted probabilities.  These should be generated from the same component models used to fit the \code{obj}
}
  \item{y.out}{
Observed outcomes for the out-of-sample dataset
}
}
\details{
This utility function is to be applied after users have fit an \code{\link{Ensemble.logit}} model.  Using new predictions from each of the component forecasts, it generates ensemble forecasts.  \cr
\cr
In addition, the function calculates fit statistics to compare the out-of-sample predictions from component models and the ensemble forecast. Setting \code{make.plot=TRUE} creates a separation plot for each component and the EBMA model. \cr
}
\value{
  \item{W}{The posterior weights assigned to each model}
  \item{pred.prob}{The posterior predicted probabilities for all n out-of-sample observations generated by the ensemble model}
  \item{model.params}{The bias-reduction constant and predictor coeficients estimated for each of the k models}
  \item{exp}{The exponential shrinkage term specified by the user}
  \item{model.names}{A string vector containing the name of each component model}
  \item{model.pps}{A n by k matrix containing the predicted probabilities from each component model (the same matrix as \code{newdata} above)}
  \item{bma.brier}{The brier score for the out-of-sample EBMA model}
  \item{models.brier}{The brier score for the component model forecasts}
  \item{bma.perc}{The percent of observations correctly predicted by the out-of-sample EBMA model}
  \item{models.perc}{The percent of observations correctly predicted by the component model forecasts}
  \item{bma.pre}{The proportional reduction in error for the out-of-sample EBMA model, where the baseline model always predicts zero.}
  \item{models.pre}{The proportional reduction in error for the component models, where the baseline model always predicts zero.}
  \item{bma.auc}{The area under the ROC curve for the out-of-sample EBMA model}
  \item{models.auc}{The area under the ROC curve for the component model forecasts}
  \item{bma.sens}{The sensitivity statistic for the out-of-sample EBMA model}
  \item{models.sens}{The sensitivity statistic for the component models}
  \item{bma.spec}{The specificity statistic for the out-of-sample EBMA model}
  \item{models.spec}{The specificity statistic for the component models}
  \item{dv}{The actual observed values for the outcome variable for the out-of-sample dataset (the same as \code{y.out} above)}
  \item{n}{The number of complete observations in the out-of-sample dataset}
}
\references{
Ward, M. D., J. M. Montgomery and F. M. Hollenbach. 2011 Improving Conflict Predictions Using Ensemble Bayesian Model Averaging.  Paper presented at the 2011 meeting of the International Studies Association in Montreal, Quebec.\cr
\cr
Raftery, A. E., T. Gneiting, F. Balabdaoui and M. Polakowski. 2005 Using Bayesian Model Averaging to calibrate forecast ensembles, \emph{Monthly Weather Review} \bold{133}:1155--1174.\cr
\cr
Sloughter, J. M., A. E. Raftery, T. Gneiting and C. Fraley. 2007 Probabilistic quantitative precipitation forecasting using Bayesian model averaging, \emph{Monthly Weather Review} \bold{135}:3209--3220.\cr
\cr
Fraley, C., A. E. Raftery, T. Gneiting, 2010 Calibrating Multi-Model Forecast Ensembles with Exchangeable and Missing Members using Bayesian Model Averaging, \emph{Monthly Weather Review} \bold{138}:190--202.\cr
\cr
Sloughter, J. M., T. Gneiting and A. E. Raftery, 2010 Probabilistic wind speed forecasting using ensembles and Bayesian model averaging, \emph{Journal of the American Statistical Association}, \bold{105}:25--35.
}
\author{
Michael D. Ward and Jacob Montgomery <jacob.montgomery@duke.edu>
}
\note{
PLEASE NOTE: This function does not currently handle missing data, but does not provide errors.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{EBMAforecast}}, \code{\link{Ensemble.logit}}, \code{\link{Insample}}, \code{\link{Outsample}}
}
\examples{
demo(EBMAforecast)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{BMA}
